Split backend API into projects, tags and hall of fame, rename `searchClient` => `api`
