+++ apps/bestofjs-nextjs/src/app/api/og/projects/[slug]/route.tsx
@@ -1,4 +1,5 @@
 import { formatNumber } from "@/helpers/numbers";
+import { api } from "@/server/api";
 import {
   Box,
   ProjectLogo,
@@ -6,7 +7,6 @@ import {
   generateImageResponse,
   mutedColor,
 } from "@/app/api/og/og-utils";
-import { searchClient } from "@/app/backend";
 
 import { ImageLayout } from "../../og-image-layout";
 
@@ -14,7 +14,7 @@ export const runtime = "edge";
 
 type Context = { params: { slug: string } };
 export async function GET(_req: Request, { params: { slug } }: Context) {
-  const project = await searchClient.getProjectBySlug(slug);
+  const project = await api.projects.getProjectBySlug(slug);
   if (!project)
     return generateImageResponse(
       <ImageLayout>
