+++ apps/bestofjs-nextjs/src/app/timeline/timeline-projects.ts
@@ -1,0 +1,157 @@
+export const timelineProjects = [
+  {
+    slug: "jquery",
+    date: "2006-08-26",
+    comments: [
+      "Created by John Resig and released in 2006, jQuery won the battle of the Ajax frameworks against libraries like Prototype.js, Dojo or YUI.",
+      "Became the de-facto standard library included in almost any web page, even today.",
+      "A beautiful piece of software at the origin of a vast ecosystem of 'plugins'.",
+      "Did a great job providing a unified API to handle browser inconsistencies.",
+    ],
+  },
+  {
+    date: "2009-05-27",
+    slug: "nodejs",
+    comments: [
+      "Created by Ryan Dahl on top of JavaScript V8 engine, Node.js brought JavaScript to the server-side.",
+      "Power APIs, command line tools, IoT devices and all the modern web development tooling.",
+    ],
+  },
+  {
+    slug: "express",
+    comments: [
+      "Almost as old as Node.js, Express is still considered as the standard web framework for Node.js.",
+      "Has a deep ecosystem of middlewares because of its minimalist approach.",
+      "More modern alternatives include Nest, Fastify or Next.js.",
+    ],
+  },
+  {
+    slug: "d3",
+    comments: [
+      "The classic library for data visualization created by Mike Bostok, still relevant after almost 10 years.",
+      "A lot of charting libraries are built on top of its powerful primitives.",
+    ],
+  },
+  {
+    slug: "angular-1",
+    comments: [
+      "In 2010, Angular from Google started to disrupt the way developers build applications by providing a full framework that runs in the browser.",
+      "Instead of enhancing the DOM generated by the server, it moves the control of the page to the client side, leading to the Single Page Application era.",
+      "Angular 2 was released as a separate project in 2014.",
+    ],
+  },
+  {
+    slug: "ember",
+    comments: [
+      "A UI framework that favors strong conventions to make web developers more productive.",
+      "Had an important role in the modernization of JavaScript.",
+    ],
+  },
+  {
+    slug: "bootstrap",
+    comments: [
+      "Open sourced by Twitter in 2011, Bootstrap is a set of JavaScript and CSS elements to add common widgets to web pages.",
+      "Shines thanks to a neat documentation, a consistent developer experience and responsive design features smartly designed.",
+      "Helped developers make web pages mobile-friendly just when browsing the web on a smartphone became the norm.",
+    ],
+  },
+  {
+    slug: "webpack",
+    comments: [
+      "A key part of the front-end tooling, Webpack is able to generate JavaScript bundles from any kind of asset.",
+      "Still used in 2020 in projects like Create React App or Next.js, the 5th major version is coming soon.",
+      "Parcel and Rollup are 2 of the most popular alternatives.",
+    ],
+  },
+  {
+    slug: "typescript",
+    date: "2012-10-01",
+    comments: [
+      "A super set of JavaScript that provides static types safety to JavaScript.",
+      "Won the battle of the transpilers against its rival Flow to become almost a standard of front-end and back-end development.",
+      "`DefinitelyTyped` project provides additional types for more than 7000 packages!",
+    ],
+  },
+  {
+    slug: "electron",
+    comments: [
+      "A solution to build desktop applications using web technologies.",
+      "Used by popular desktop applications such as VS Code, Slack...",
+    ],
+  },
+  {
+    slug: "react",
+    comments: [
+      "React was open-sourced by Facebook to create dynamic and fast UI.",
+      "Not a full-framework like Angular as it handles only the view layer.",
+      "Its component approach and the JSX syntax became very popular and inspired a lot of projects.",
+    ],
+  },
+  {
+    slug: "vuejs-2",
+    comments: [
+      "Released only 2 months after React, Vue.js takes a different approach to handle reactivity.",
+      "Often considered as more beginner-friendly.",
+      "Not backed by a huge company like React with Facebook or Angular with Google but created by Evan You.",
+      "Strong community, especially in China.",
+      "The version 3 provides huge improvements including a better TypeScript support and the Composition API.",
+    ],
+  },
+  {
+    slug: "babel",
+    comments: [
+      "First called `6to5`, the compiler Babel helped make the ES6 version of JavaScript popular and was a key part of React success.",
+      "Let developers write code using the latest features of JavaScript without having to worry too much about the browser support.",
+      "Included as a dependency by a lot of projects.",
+    ],
+  },
+  {
+    slug: "vs-code",
+    comments: [
+      "The most popular IDE to code JavaScript and TypeScript applications, built on top of Electron.",
+      "Atom from GitHub, launched in 2012, was the pioneer of the concept: a code editor, built with web technologies, extensible by installing plugins.",
+      "Shines with its TypeScript support... both projects being from Microsoft.",
+    ],
+  },
+  {
+    slug: "react-native",
+    comments: [
+      "React Native lets developers build mobile applications for iOS or Android leveraging their knowledge of the React framework.",
+    ],
+  },
+  {
+    slug: "nextjs",
+    comments: [
+      "Became the most popular framework to build full-stack JavaScript applications (applications that run on both front and back ends).",
+      "Uses React in the front-end and provides Server-Side Rendering features to deliver the best user-experience.",
+    ],
+  },
+  {
+    slug: "puppeteer",
+    comments: [
+      "A Node.js library to control a headless Chrome browser to to do things like browser testing or web scrapping.",
+    ],
+  },
+  {
+    slug: "deno",
+    comments: [
+      "9 years after the release of Node.js, Ryan Dahl strikes again with a new server-side JavaScript run-time.",
+      "Aims to solves several issues related to Node.js.",
+      "No centralized package manager, any JavaScript dependency can be loaded from any URL.",
+      "One of the most popular projects in 2020.",
+    ],
+  },
+  {
+    slug: "snowpack",
+    comments: [
+      "A building tool that takes advantage of the latest changes about JavaScript modules to provide a great developer experience.",
+    ],
+  },
+  {
+    slug: "rome",
+    comments: [
+      "Created by Seb McKenzie, Rome aims to provide a single dependency to handle all the front-end tooling (linting, testing, building...).",
+      "Will it be the success story of 2020?",
+    ],
+  },
+];
